DJANGO/14_Test/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
coverage==7.0.3
dj-rest-auth==2.2.5
Django==4.1.4
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==22.0
psycopg2-binary==2.9.5
python-decouple==3.6
pytz==2022.7
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
